refactor: MainViewController 하나만 사용하도록 변경
