--- OpenMarket/OpenMarket/Controller/ProductsCollectionViewController.swift
@@ -1,150 +1,0 @@
-import UIKit
-
-class ProductsCollectionViewController: UICollectionViewController {
-    private let loadingActivityIndicator = UIActivityIndicatorView()
-    private var pageInformation: ProductsList?
-    private var products = [Product]()
-    private var jsonParser: JSONParser?
-    private var networkTask: NetworkTask?
-    
-    convenience init?(coder: NSCoder, jsonParser: JSONParser, networkTask: NetworkTask) {
-        self.init(coder: coder)
-        self.jsonParser = jsonParser
-        self.networkTask = networkTask
-    }
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        startActivityIndicator()
-        loadProductsList(pageNumber: 1)
-    }
-    
-    private func startActivityIndicator() {
-        view.addSubview(loadingActivityIndicator)
-        
-        loadingActivityIndicator.translatesAutoresizingMaskIntoConstraints = false
-        loadingActivityIndicator.centerYAnchor.constraint(
-            equalTo: view.centerYAnchor
-        ).isActive = true
-        loadingActivityIndicator.centerXAnchor.constraint(
-            equalTo: view.centerXAnchor
-        ).isActive = true
-        loadingActivityIndicator.startAnimating()
-    }
-    
-    private func loadProductsList(pageNumber: Int) {
-        networkTask?.requestProductList(pageNumber: pageNumber, itemsPerPage: 20) { result in
-            switch result {
-            case .success(let data):
-                guard let productsList: ProductsList = try? self.jsonParser?.decode(
-                    from: data
-                ) else { return }
-                self.pageInformation = productsList
-                self.products.append(contentsOf: productsList.pages)
-                DispatchQueue.main.async {
-                    self.loadingActivityIndicator.stopAnimating()
-                    self.collectionView.reloadData()
-                }
-            case .failure(let error):
-                self.showAlert(
-                    title: "Network error",
-                    message: "데이터를 불러오지 못했습니다.\n\(error.localizedDescription)"
-                )
-                self.loadingActivityIndicator.stopAnimating()
-            }
-        }
-    }
-    
-    private func setupCellImage(
-        for cell: ProductsCollectionViewCell,
-        from url: URL,
-        indexPath: IndexPath,
-        collectionView: UICollectionView
-    ) {
-        cell.setup(imageView: nil)
-        networkTask?.downloadImage(from: url) { result in
-            switch result {
-            case .success(let data):
-                guard let image = UIImage(data: data) else { return }
-                DispatchQueue.main.async {
-                    guard indexPath == collectionView.indexPath(for: cell) else { return }
-                    cell.setup(imageView: image)
-                }
-            case .failure(let error):
-                self.showAlert(
-                    title: "Network error",
-                    message: "데이터를 불러오지 못했습니다.\n\(error.localizedDescription)"
-                )
-            }
-        }
-    }
-    
-    // MARK: - Collection view Data source
-    
-    override func numberOfSections(in collectionView: UICollectionView) -> Int {
-        return 1
-    }
-    
-    override func collectionView(
-        _ collectionView: UICollectionView,
-        numberOfItemsInSection section: Int
-    ) -> Int {
-        return products.count
-    }
-    
-    override func collectionView(
-        _ collectionView: UICollectionView,
-        cellForItemAt indexPath: IndexPath
-    ) -> UICollectionViewCell {
-        let product = products[indexPath.item]
-        let reuseIdentifier = ProductsCollectionViewCell.reuseIdentifier
-        let cell = collectionView.dequeueReusableCell(
-            withReuseIdentifier: reuseIdentifier,
-            for: indexPath
-        )
-        guard let cell = cell as? ProductsCollectionViewCell,
-              let url = URL(string: product.thumbnail) else {
-                  return cell
-              }
-        cell.setup(
-            titleLabel: product.attributedTitle,
-            priceLabel: product.attributedPrice,
-            stockLabel: product.attributedStock
-        )
-        setupCellImage(for: cell, from: url, indexPath: indexPath, collectionView: collectionView)
-        return cell
-    }
-    
-    // MARK: - Collection view delegate
-    
-    override func collectionView(
-        _ collectionView: UICollectionView,
-        willDisplay cell: UICollectionViewCell,
-        forItemAt indexPath: IndexPath
-    ) {
-        if indexPath.item == products.count - 1,
-           pageInformation?.hasNext == true,
-           let num = pageInformation?.pageNumber {
-            loadProductsList(pageNumber: num + 1)
-        }
-    }
-}
-
-extension ProductsCollectionViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(
-        _ collectionView: UICollectionView,
-        layout collectionViewLayout: UICollectionViewLayout,
-        sizeForItemAt indexPath: IndexPath
-    ) -> CGSize {
-        let frameWidth = collectionView.frameLayoutGuide.layoutFrame.width
-        let frameHeight = collectionView.frameLayoutGuide.layoutFrame.height
-        let shortLength = frameWidth < frameHeight ? frameWidth : frameHeight
-        let cellWidth = shortLength / 2 - 15
-        return CGSize(width: cellWidth, height: cellWidth * 1.5)
-    }
-}
-
-extension ProductsCollectionViewController {
-    static let gridViewStoryboardName = "ProductsCollectionView"
-    static let gridViewControllerIdentifier = "GridViewController"
-}
