+++ OpenMarket/OpenMarket/Extension/UITableView+extension.swift
@@ -1,0 +1,10 @@
+import UIKit
+
+extension UITableView: ProductView {
+    func indexPath(for cell: ProductCell) -> IndexPath? {
+        guard let tableViewCell = cell as? UITableViewCell else {
+            return nil
+        }
+        return indexPath(for: tableViewCell)
+    }
+}
