+++ OpenMarket/OpenMarket/Protocol/ProductCell.swift
@@ -1,0 +1,10 @@
+import UIKit
+
+protocol ProductCell {
+    func setup(
+        titleLabel: NSAttributedString,
+        priceLabel: NSAttributedString,
+        stockLabel: NSAttributedString
+    )
+    func setup(imageView: UIImage?)
+}
