+++ OpenMarket/OpenMarket/Protocol/ProductView.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+protocol ProductView {
+    func indexPath(for cell: ProductCell) -> IndexPath?
+    func reloadData()
+}
