+++ OpenMarket/OpenMarket/View/ProductsCollectionViewCell.xib
@@ -1,0 +1,67 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="19455" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19454"/>
+        <capability name="collection view cell content view" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="productCell" id="e9K-DD-NmS" customClass="ProductsCollectionViewCell" customModule="OpenMarket" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="185" height="190"/>
+            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+            <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="BjF-FT-nG7">
+                <rect key="frame" x="0.0" y="0.0" width="185" height="190"/>
+                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                <subviews>
+                    <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="center" spacing="8" translatesAutoresizingMaskIntoConstraints="NO" id="5QG-KW-hbC">
+                        <rect key="frame" x="10" y="10" width="165" height="170"/>
+                        <subviews>
+                            <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="249" horizontalCompressionResistancePriority="250" verticalCompressionResistancePriority="250" translatesAutoresizingMaskIntoConstraints="NO" id="v75-JS-qI7">
+                                <rect key="frame" x="36.5" y="0.0" width="92.5" height="84.5"/>
+                            </imageView>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="title" textAlignment="center" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="ZA1-HP-nNa">
+                                <rect key="frame" x="68.5" y="92.5" width="28.5" height="20.5"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalCompressionResistancePriority="251" text="price" textAlignment="center" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="lIm-dQ-8tX">
+                                <rect key="frame" x="63.5" y="121" width="38.5" height="20.5"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="stock" textAlignment="center" lineBreakMode="wordWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="WE0-no-8hZ">
+                                <rect key="frame" x="62" y="149.5" width="41.5" height="20.5"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                        </subviews>
+                        <constraints>
+                            <constraint firstItem="v75-JS-qI7" firstAttribute="height" relation="lessThanOrEqual" secondItem="5QG-KW-hbC" secondAttribute="height" multiplier="0.6" id="QIU-oQ-RC9"/>
+                        </constraints>
+                    </stackView>
+                </subviews>
+                <constraints>
+                    <constraint firstAttribute="bottom" secondItem="5QG-KW-hbC" secondAttribute="bottom" constant="10" id="8Nc-r8-iAo"/>
+                    <constraint firstItem="5QG-KW-hbC" firstAttribute="leading" secondItem="BjF-FT-nG7" secondAttribute="leading" constant="10" id="NOl-0o-ujk"/>
+                    <constraint firstAttribute="trailing" secondItem="5QG-KW-hbC" secondAttribute="trailing" constant="10" id="b8E-kO-8JQ"/>
+                    <constraint firstItem="5QG-KW-hbC" firstAttribute="top" secondItem="BjF-FT-nG7" secondAttribute="top" constant="10" id="mLn-Uq-ruY"/>
+                </constraints>
+            </collectionViewCellContentView>
+            <size key="customSize" width="185" height="190"/>
+            <connections>
+                <outlet property="productImageView" destination="v75-JS-qI7" id="UnX-cq-6T2"/>
+                <outlet property="productPriceLabel" destination="lIm-dQ-8tX" id="08u-tn-MuP"/>
+                <outlet property="productStockLabel" destination="WE0-no-8hZ" id="KZf-56-4LJ"/>
+                <outlet property="productTitleLabel" destination="ZA1-HP-nNa" id="5f8-2V-6Hh"/>
+            </connections>
+            <point key="canvasLocation" x="1014" y="-500"/>
+        </collectionViewCell>
+    </objects>
+</document>
